backend/requirements.txt
pandas
pandas_ta
mplfinance
matplotlib
langchain
langchain-google-genai
langgraph
fastapi
uvicorn
httpx
apscheduler
openalgo
python-dotenv
duckduckgo-search
langchain-community
filetype

google-genai
